move export car dialog
